feat: add support for `set`s and `dict`s for rpc api
