+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - fix: vscode binary now needs to be instructed about the location of the binary with `version use` sub-command - closes #217
 - fix: image size being too big for the lite version of the raspberry os
 - fix: restart avahi-daemon after the hostname is set in `ubo-system`
+- feat: add support for `set`s and `dict`s for rpc api
 
 ## Version 1.2.2
