+++ ubo_app/rpc/generator/__init__.py
@@ -1,0 +1,1 @@
+"""Generate proto files from actions and events defined in Python files."""
